@@ -14,40 +14,40 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.16",
+    "@angular/animations": "^8.2.14",
     "@angular/cdk": "^7.3.3",
-    "@angular/common": "~7.1.0",
-    "@angular/compiler": "~7.1.0",
-    "@angular/core": "~7.1.0",
-    "@angular/forms": "~7.1.0",
+    "@angular/common": "~8.2.14",
+    "@angular/compiler": "~8.2.14",
+    "@angular/core": "~8.2.14",
+    "@angular/forms": "~8.2.14",
     "@angular/material": "^7.3.3",
-    "@angular/platform-browser": "~7.1.0",
-    "@angular/platform-browser-dynamic": "~7.1.0",
-    "@angular/router": "~7.1.0",
+    "@angular/platform-browser": "~8.2.14",
+    "@angular/platform-browser-dynamic": "~8.2.14",
+    "@angular/router": "~8.2.14",
     "@auth0/angular-jwt": "^2.1.2",
     "@fortawesome/angular-fontawesome": "^0.3.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.28",
     "@fortawesome/free-solid-svg-icons": "^5.13.0",
     "@ngx-translate/core": "^12.1.2",
     "@ngx-translate/http-loader": "^4.0.0",
-    "@swimlane/ngx-charts": "^10.1.0",
     "angular-progress-bar": "^1.0.11",
     "core-js": "^2.6.11",
     "http-status-codes": "^1.4.0",
     "jwt-decode": "^2.2.0",
     "lodash": "^4.17.15",
     "moment": "^2.24.0",
-    "ngx-socket-io": "^3.0.1",
     "reset-css": "^4.0.1",
-    "rxjs": "~6.3.3",
+    "rxjs": "~6.6.3",
     "tslib": "^1.11.1",
-    "zone.js": "~0.8.26"
+    "ngx-socket-io": "^3.0.1",
+    "@swimlane/ngx-charts": "^13.0.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.13.10",
-    "@angular/cli": "~7.1.2",
-    "@angular/compiler-cli": "^7.2.16",
-    "@angular/language-service": "~7.1.0",
+    "@angular/cli": "~8.3.29",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "~8.2.14",
     "@biesbjerg/ngx-translate-extract": "^4.2.0",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "^2.0.8",
@@ -68,6 +68,6 @@
     "stylelint-scss": "^3.16.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.5.3"
   }
 }
