@@ -22,8 +22,8 @@ export class ProfilesComponent implements OnInit {
   public faTrashAlt = faTrashAlt;
   public faDoorOpen = faDoorOpen;
 
-  @ViewChild(MatSort) sort: MatSort;
-  @ViewChild(MatPaginator) paginator: MatPaginator;
+  @ViewChild(MatSort, { static: false }) sort: MatSort;
+  @ViewChild(MatPaginator, { static: false }) paginator: MatPaginator;
 
   constructor(
     private _cdr: ChangeDetectorRef,
